build, bugfix: compile warning on Windows.

+++ fibjs/include/HttpCollection.h
@@ -61,7 +61,7 @@ public:
 
     result_t first(exlib::string name, exlib::string& retVal)
     {
-        int32_t i;
+        size_t i;
 
         for (i = 0; i < m_count; i++) {
             pair& _pair = m_map[i];
@@ -78,7 +78,7 @@ public:
     result_t all(exlib::string name, obj_ptr<NArray>& retVal)
     {
         obj_ptr<NArray> list = new NArray();
-        int32_t i;
+        size_t i;
 
         for (i = 0; i < m_count; i++) {
             pair& _pair = m_map[i];
@@ -94,7 +94,7 @@ public:
     result_t all(obj_ptr<NObject>& retVal)
     {
         obj_ptr<NObject> map = new NObject(true);
-        int32_t i;
+        size_t i;
 
         for (i = 0; i < m_count; i++) {
             pair& _pair = m_map[i];
@@ -114,7 +114,7 @@ public:
 private:
     typedef std::pair<exlib::string, exlib::string> pair;
     std::vector<pair> m_map;
-    int32_t m_count;
+    size_t m_count;
 };
 
 } /* namespace fibjs */
